+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased]
+
+### Performance
+
+* context: Do not reinitialize config when adding a single index.
+
 ## 5.1.2 - 2022-05-24
 
 ### Added